% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_esquisse_module.R
\name{example_esquisse_module}
\alias{example_esquisse_module}
\title{An example \code{teal} module}
\usage{
example_esquisse_module(
  label = "example teal module",
  datanames = "all",
  transformers = list()
)
}
\value{
A \code{teal} module which can be included in the \code{modules} argument to \code{\link[=init]{init()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = example_module()
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
